Update layer_factory_impl.h
when i use the caffe_convert tools to convert my caffe model to tiny-dnn,  an error occured, becasue a convolutional layer which shape is 2 x 3 is not supported. with the Error message is :"unsupported kernel shape" in line 97. So i avoid the check in "get_kernel_size_2d" and directly to get the kernel_h and kernal_w  from "conv_param".